// src/utils/enums.ts
export enum Range {
  HOUR = "Hour",
  MINUTE = "Minute",
  DAY = "Day",
}

export enum Metric {
  TEMPERATURE = "Temperature",
  HUMIDITY = "Humidity",
  PRESSURE = "Pressure",
  LOW = "Low",
  HIGH = "High",
  PRECIPITATION = "Precipitation",
  VISIBILITY = "Visibility",
  WIND_SPEED = "Wind Speed",
  FEELS_LIKE = "Feels Like",
}